pyproject.toml
[project]
name = "hireai-restapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.13",
    "llama-index>=0.12.43",
    "llama-index-embeddings-fastembed>=0.3.2",
    "llama-index-embeddings-jinaai>=0.4.0",
    "llama-index-llms-openai>=0.4.7",
    "llama-index-postprocessor-jinaai-rerank>=0.3.0",
    "llama-index-readers-smart-pdf-loader>=0.3.0",
    "llama-index-vector-stores-chroma>=0.4.2",
    "llama-index-vector-stores-qdrant>=0.6.0",
    "llmsherpa>=0.1.4",
    "protobuf>=3.20,<3.21",
    "pydantic-settings>=2.9.1",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.14.3",
    "uvicorn>=0.34.3",
]
